feature(partitions): remove gemini partitions type (#551)
* feature(partitions): remove gemini partitions type

Due to concept being too confusing and prone to frequent breaking
in the part of the code, it has been decided that these partitions need
to go out in favour of simple solution which will hold all partitions
keys in memory (later we will utilize the resources better) and pick
them according to distribution function used

* feature(partitions): implement time bucketing system for data resurrection validation


* feature(partititons): validate deleted partitions


* feature(flags): add flags for deleted-time-buckets queries


* feature(partitions): limit the partition key size


* fix(deleted): deleted partitions just need to compare with oracle


---------

+++ pkg/distributions/distributions_test.go
@@ -30,7 +30,7 @@ func TestNew(t *testing.T) {
 
 	data := []struct {
 		dist          Distribution
-		size          int
+		size          uint64
 		seed          uint64
 		mu, sigma     float64
 		maxSameValues int
